modify and add new files

+++ TodoApp/routers/todos.py
@@ -1,9 +1,4 @@
 from fastapi import  APIRouter
-
-router = APIRouter()
-
-
-
 from fastapi import  Depends, HTTPException, Path
 from pydantic import BaseModel, Field
 from database import SessionLocal
@@ -11,6 +6,12 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from models import Todos
 from starlette import status
+
+from .auth import get_current_user
+
+router = APIRouter()
+
+
 
 
 
@@ -24,6 +25,8 @@ def get_db():
 
 db_dependency = Annotated[Session, Depends(get_db)]     
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
+
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str
@@ -32,26 +35,34 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()  
+async def read_all(user:user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    return db.query(Todos).filter(Todos.owner_id==user.get('id')).all()  
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user:user_dependency,db: db_dependency, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo list not found')
        
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request:TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency,db: db_dependency, todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = Todos(**todo_request.model_dump(), owner_id = user.get('id'))
     db.add(todo_model)
     db.commit()
     return "Successfully added the todo list item"
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency, todo_request:TodoRequest, todo_id:int  = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user:user_dependency,db:db_dependency, todo_request:TodoRequest, todo_id:int  = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo item not found")
     todo_model.title = todo_request.title
@@ -62,9 +73,11 @@ async def update_todo(db:db_dependency, todo_request:TodoRequest, todo_id:int  =
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency, todo_id:int  = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency, todo_id:int  = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo item not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
